Improve error management using WebSocketException and slf4j logger #120 #121

+++ backend/src/test/java/org/metacorp/mindbug/utils/TestUtils.java
@@ -1,14 +1,20 @@
 package org.metacorp.mindbug.utils;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.text.MessageFormat;
 import java.util.stream.Stream;
 
 /**
  * Utility class for tests
  */
 public class TestUtils {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(TestUtils.class);
 
     /**
      * Clean the directory containing log files
@@ -20,14 +26,15 @@ public class TestUtils {
 
         try (Stream<Path> stream = Files.list(logDir)) {
             stream.forEach(path -> {
-                try {
-                    Files.deleteIfExists(path);
-                } catch (IOException e) {
-                    e.printStackTrace();
+                if (!path.getFileName().toString().equals("mindbug.log")) {
+                    try {
+                        Files.deleteIfExists(path);
+                    } catch (IOException e) {
+                        String errorMessage = MessageFormat.format("Failed to clean history file {0}", path);
+                        LOGGER.warn(errorMessage, e);
+                    }
                 }
             });
         }
-
-        Files.delete(logDir);
     }
 }
